Room, Room Manager, Server, UserHandler changed
1. RoomManager.java is designed to make a new room by compatible with
Room Class' instance.
2. Room.java is designed to store the room's informations and now it has
basic variables but we have to put more for perfect room system.
3. Server.java is designed to make server socket and allow clients to
enter the server. And it process almost of methods that make
communication with clients.
4. UserHandler.java is designed to record a user that connected to the
server, and it also run by compatible with User.java
5. User.java is designed to store user information. Actually, a user's
state are managed by mySQL server, but a records of online user will be
synchronized by this class.